requirements.txt
requests
beautifulsoup4
markdownify
pymupdf
google-generativeai
openai
